wf-2: refact(weather): moving method to convert result forecast in data class
