+++ src/main/java/com/cawe/horaperfeita/infrastructure/external/interfaces/IForecast.java
@@ -1,9 +1,11 @@
 package com.cawe.horaperfeita.infrastructure.external.interfaces;
 
+import com.cawe.horaperfeita.infrastructure.external.data.WeatherData;
+
 public interface IForecast {
     IForecast addTemperature();
 
     IForecast addLatitudeAndLongitude(String latitude, String longitude);
 
-    String getWeatherForecast();
+    WeatherData getWeatherForecast();
 }
